socket/requirements.txt
bidict==0.22.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
Flask-SocketIO==5.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.29.0
urllib3==1.26.15
Werkzeug==2.3.1
